Added data augmentation scripts
Occlusion augmentation adds clutter that randomly covers the images with geometric primitives up to a certain percentage. The tile augmentation randomly permutes a set number of image tiles. This can be used to test the robustness of part-whole relations (the classifier should not classify the tiled objects correctly or have low confidence in the classification).